+++ firmware/controllers/math/engine_math.cpp
@@ -74,7 +74,7 @@ void setSingleCoilDwell() {
 /**
  * @return Spark dwell time, in milliseconds. 0 if tables are not ready.
  */
-floatms_t IgnitionState::getSparkDwell(int rpm) {
+floatms_t IgnitionState::getSparkDwell(float rpm) {
 #if EFI_ENGINE_CONTROL && EFI_SHAFT_POSITION_INPUT
 	float dwellMs;
 	if (engine->rpmCalculator.isCranking()) {
@@ -99,7 +99,7 @@ floatms_t IgnitionState::getSparkDwell(int rpm) {
 
 	if (std::isnan(dwellMs) || dwellMs <= 0) {
 		// this could happen during engine configuration reset
-		warning(ObdCode::CUSTOM_ERR_DWELL_DURATION, "invalid dwell: %.2f at rpm=%d", dwellMs, rpm);
+		warning(ObdCode::CUSTOM_ERR_DWELL_DURATION, "invalid dwell: %.2f at rpm=%.0f", dwellMs, rpm);
 		return 0;
 	}
 	return dwellMs;
