+++ unit_tests/tests/ignition_injection/test_fuel_math.cpp
@@ -145,7 +145,7 @@ TEST(AirmassModes, VeOverride) {
 	struct DummyAirmassModel : public AirmassVeModelBase {
 		DummyAirmassModel(const ValueProvider3D& veTable) : AirmassVeModelBase(veTable) {}
 
-		AirmassResult getAirmass(int rpm, bool postState) override {
+		AirmassResult getAirmass(float rpm, bool postState) override {
 			// Default load value 10, will be overriden
 			getVe(rpm, 10.0f, postState);
